LayoutTests/http/tests/websocket/tests/hybi/contentextensions/block-worker-expected.txt

CONSOLE MESSAGE: Content blocker prevented frame displaying http://127.0.0.1:8000/websocket/tests/hybi/contentextensions/block-worker.html from loading a resource from ws://localhost:8880/websocket/tests/hybi/contentextensions/resources/echo
CONSOLE MESSAGE: line 21: PASS worker onerror
